requirements.txt
streamlit
torch
langchain_huggingface
faiss-cpu
langchain_community
langchain_text_splitters
langchain_community
langchain
dotenv
langchain_ollama
langchain_openai
langchain-groq